// src/types.ts
export interface GregorianParts {
  gYear: number;
  gMonth: number;
  gDay: number;
}

export interface JalaliMathParts {
  jYear: number;
  jMonth: number;
  jDay: number;
}

export interface JalaliDateParts {
  year: number;
  month: number;
  day: number;
  weekday: number;
  hour: number;
  minute: number;
  second: number;
}